# Conflicts:
#	app/Http/Controllers/Auth/RegisterController.php

+++ app/Services/Program/ProgramTemplate.php
@@ -271,14 +271,15 @@ abstract class ProgramTemplate
         }
 
         // material_id => sometimes 이기 때문.
-        $data['material_id'] = isset($data['material_id']) ?: null;
+        $data['material_id'] = isset($data['material_id']) ? $data['material_id'] : null;
 
         if ($program->material_id != $data['material_id']) {
             // 변경 있음.
             $fileService = new ProgramMaterial($program);
-
-            //기존 파일 삭제
-            $fileService->deleteFile();
+            if ($program->material()->exists()) {
+                //기존 파일 삭제
+                $fileService->deleteFile();
+            }
 
             if ($data['material_id'] !== null) {
                 // 새 파일 생성
